# backend/offertracker-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OfferTracker API",
    "description": "求职管理系统后端API",
    "version": "1.0.0",
    "contact": {
      "name": "OfferTracker Team",
      "url": "https://github.com/offertracker"
    }
  },
  "servers": [
    {
      "url": "https://offertracker-backend.1544718463.workers.dev",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:8787",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/resumes": {
      "get": {
        "summary": "获取简历列表",
        "description": "获取当前用户的所有简历",
        "tags": ["简历管理"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取简历列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resume"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建简历",
        "description": "创建新的简历记录并获取上传URL",
        "tags": ["简历管理"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["resume_name"],
                "properties": {
                  "resume_name": {
                    "type": "string",
                    "example": "前端开发简历"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "简历创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "resume-id"
                    },
                    "file_url": {
                      "type": "string",
                      "example": "resumes/user-id/resume-id.pdf"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/{id}": {
      "patch": {
        "summary": "更新简历信息",
        "description": "更新简历的名称信息",
        "tags": ["简历管理"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resume_name": {
                    "type": "string",
                    "example": "更新后的简历名称"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "summary": "删除简历",
        "description": "删除指定的简历",
        "tags": ["简历管理"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/api/resumes/{id}/file": {
      "put": {
        "summary": "上传简历文件",
        "description": "上传简历文件到指定的简历记录",
        "tags": ["简历管理"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文件上传成功"
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "简历记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/{id}/download": {
      "get": {
        "summary": "获取简历下载信息",
        "description": "获取简历文件的下载信息",
        "tags": ["简历管理"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取下载信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "file_path": {
                      "type": "string",
                      "example": "resumes/user-id/resume-id.pdf"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "错误信息"
          }
        }
      },
...
      "Resume": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "resume-id"
          },
          "user_id": {
            "type": "string",
            "example": "user-id"
          },
          "resume_name": {
            "type": "string",
            "example": "前端开发简历"
          },
          "file_url": {
            "type": "string",
            "example": "resumes/user-id/resume-id.pdf"
          },
          "created_at": {
            "type": "integer",
            "example": 1234567890
          },
          "updated_at": {
            "type": "integer",
            "example": 1234567890
          }
        }
      },
...
    }
  }
}